+++ src/main/java/Main.java
@@ -1,6 +1,6 @@
 public class Main {
     public static void main(String[] args) {
-        Server server = new Server(64);
+        var server = new Server(64);
         server.start(9999);
     }
 }
